* Restore middle-mouse and right-mouse menu behaviours
Note that up and including to 1.5, it was possible to open
the menu's target page in a new window using middle-click
or right-click/'Open link in new window'. In 1.6, that's
not possible due to the use of the dijit/Tree-based menu
which doesn't allow catching all variants of events.

Using this approach, I'm able to add back the old behaviour
while still retaining the nice tree approach based on Vue.